requirements.txt
# python 3.11.9

Flask
requests
python-dotenv
openai